Adding singleton class to store information from the BRB settings tree
There is new singleton class BrbSettingsTree, which is meant to
store information like the online calculated baseline and HV setpoints;
this information is stored in a special TTree in the ROOT file.
This tree is open by calling a method in wrapper and the
interesting settings are then saved.  These baseline and HV values
can then be used in the mPMT pulse finding and pulse fitting.